+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.7] - 2025-06-25
 ### Added
 - `Message` abstract class
 - `Session` class `_validate_extract_json` method
@@ -105,7 +106,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `PresetPromptTemplate` class
 
 
-[Unreleased]: https://github.com/openscilab/memor/compare/v0.6...dev
+[Unreleased]: https://github.com/openscilab/memor/compare/v0.7...dev
+[0.7]: https://github.com/openscilab/memor/compare/v0.6...v0.7
 [0.6]: https://github.com/openscilab/memor/compare/v0.5...v0.6
 [0.5]: https://github.com/openscilab/memor/compare/v0.4...v0.5
 [0.4]: https://github.com/openscilab/memor/compare/v0.3...v0.4